feat/refactor: use of customtkinter, separated labels, root Windows and select validation.
Now customtkinter library is used along with few tkinter methods to get a better and 'modern' view.

Labels were divided to get a differentiation color for the user's use choice.

The root window was passed as parameter to get everything in the same window and avoid creating a new one.

Finally, a valdiation for the selects was added before continuing with the following next steps.